Account for single quotes in search strings
According to the SQlite docs:

Strings that include any other characters must be quoted. Characters
that are not currently allowed in barewords, are not quote characters
and do not currently serve any special purpose in FTS5 query expressions
may at some point in the future be allowed in barewords or used to
implement new query functionality. This means that queries that are
currently syntax errors because they include such a character outside of
a quoted string may be interpreted differently by some future version of
FTS5.